+++ cx-portal/src/features/appManagement/actions.ts
@@ -19,9 +19,12 @@
  ********************************************************************************/
 
 import { createAction } from '@reduxjs/toolkit'
-import { name, SearchInputState } from './types'
+import { name, SearchInputState, AppStatusDataState } from './types'
 
 const setSearchInput = createAction<SearchInputState>(`${name}/setSearchOpen`)
 const setAppId = createAction<string>(`${name}/setApplicationId`)
+const setAppStatus = createAction<AppStatusDataState>(
+  `${name}/setAppStatusData`
+)
 
-export { setSearchInput, setAppId }
+export { setSearchInput, setAppId, setAppStatus }